refact: split lib and executable
